+++ src/ws_editdb.cpp
@@ -29,6 +29,7 @@
  */
 
 #include <ctime>
+#include <iomanip>
 #include <memory>
 #include <mutex>
 #include <optional>